Create Wipro ques. 1
Find String Code

Crazy Zak has designed the below steps which can be applied on any given string (sentence) to produce a number

STEP1. in each word, find the Sum of the Difference between the first letter and the last letter, second letter and the penultimate letter and so on till the center of

the word

STEP2. Concatenate the sums of each word to form the result

For example

vir the given string is "WORLD WIDE WEB

STEPI. In each word. find the Sum of the Difference between the first letter and the last letter second letter and the penultimate letter, and so on till the center of the word

WORLO W-D]10--R- 12341512 [16] (19+16)-(40] WIDE (W-E•-01 (23-51+19-41 [18-15] (23) WEB (W-B)-[123-21-[5] [21] 15)-126]